+++ src/main/native/shaders/optix_shaders.cu
@@ -16,6 +16,7 @@ using namespace RayTracingConstants;
 #include "hit_triangle.cu"
 #include "hit_cylinder.cu"
 #include "hit_cone.cu"
+#include "hit_curve.cu"
 #include "hit_plane.cu"
 #include "hit_menger4d.cu"
 #include "hit_sierpinski4d.cu"